Dernier commit

+++ src/main/java/fr/unice/miage/sd/tinydfs/nodes/MasterImpl.java
@@ -26,8 +26,12 @@ public class MasterImpl extends UnicastRemoteObject implements Master {
 	private Slave rightSlave;
 	private Slave leftSlave;
 	private boolean isBuilded;
-	private HashMap<String, List<Thread>> fileLocked; // Un hashmap contiens la liste des Saves en cours
-//Avoir une liste permet d'avoir plusieur Saves sur le même filename en meme temps .
+	private HashMap<String, List<Thread>> fileLocked; // Un hashmap contiens la
+														// liste des Saves en
+														// cours
+	// Avoir une liste permet d'avoir plusieur Saves sur le même filename en
+	// meme temps .
+
 	public MasterImpl(String dfsRootFolder, int nbSlave) throws RemoteException, WrongNbSlaveException {
 		super();
 		if ((nbSlave + 2 & nbSlave + 1) != 0) {
@@ -35,21 +39,19 @@ public class MasterImpl extends UnicastRemoteObject implements Master {
 		}
 		this.dfsRootFolder = dfsRootFolder;
 		File dfsFileRootFolder = new File(dfsRootFolder);
-		if(!dfsFileRootFolder.exists()){
+		if (!dfsFileRootFolder.exists()) {
 			dfsFileRootFolder.mkdir();
 			System.out.println("Création dossier " + dfsFileRootFolder.getName());
-		}
-		else
-		{
+		} else {
 			File[] oldFilesSlave = dfsFileRootFolder.listFiles();
-			for (File oldFile:oldFilesSlave) {
-				System.out.println("Suppression du fichier "  + oldFile.getName());
+			for (File oldFile : oldFilesSlave) {
+				System.out.println("Suppression du fichier " + oldFile.getName());
 				oldFile.delete();
 			}
 
 		}
 		this.isBuilded = false;
-		this.nbSlave=nbSlave;
+		this.nbSlave = nbSlave;
 		this.slave = new Slave[nbSlave];
 		this.rightSlave = null;
 		this.leftSlave = null;
@@ -87,22 +89,25 @@ public class MasterImpl extends UnicastRemoteObject implements Master {
 			public void run() {
 				System.out.println("Thread running");
 				List<byte[]> divideFile = getMultipleByteArray2(fileContent);
-				List<byte[]> forLeftSlave = new ArrayList<byte[]>(divideFile.subList(0, divideFile.size()/2));
-				List<byte[]> forRightSlave = new ArrayList<byte[]>(divideFile.subList(divideFile.size()/2, divideFile.size())) ;
+				List<byte[]> forLeftSlave = new ArrayList<byte[]>(divideFile.subList(0, divideFile.size() / 2));
+				List<byte[]> forRightSlave = new ArrayList<byte[]>(
+						divideFile.subList(divideFile.size() / 2, divideFile.size()));
 				try {
 					leftSlave.subSave(filename, forLeftSlave);
 					rightSlave.subSave(filename, forRightSlave);
 				} catch (RemoteException e) {
 					e.printStackTrace();
 				}
-				fileLocked.get(filename).remove(this); // Une fois le thread terminé, il se retire lui même de la liste
+				fileLocked.get(filename).remove(this); // Une fois le thread
+														// terminé, il se retire
+														// lui même de la liste
 			}
-		}); 
-		//Sauvegarde des threads dans la liste correspondant à la Clé 
-		if(!fileLocked.containsKey(filename)){
+		});
+		// Sauvegarde des threads dans la liste correspondant à la Clé
+		if (!fileLocked.containsKey(filename)) {
 			fileLocked.put(filename, new ArrayList<Thread>());
 		}
-		fileLocked.get(filename).add(threadRetrieve); 
+		fileLocked.get(filename).add(threadRetrieve);
 		threadRetrieve.start();
 		///////////////////////////////
 	}
@@ -110,9 +115,12 @@ public class MasterImpl extends UnicastRemoteObject implements Master {
 	@Override
 	public File retrieveFile(String filename) throws RemoteException {
 		byte[] b = retrieveBytes(filename);
+		if(b==null) {
+			return null;
+		}
 		File res = new File(dfsRootFolder + File.separator + filename);
 		try {
-			if(!res.exists()) {
+			if (!res.exists()) {
 				res.createNewFile();
 			} else {
 				res.delete();
@@ -133,14 +141,27 @@ public class MasterImpl extends UnicastRemoteObject implements Master {
 			buildBinaryTree();
 			isBuilded = true;
 		}
-		for (Thread retrieve : fileLocked.get(filename)) { // Si un thread ou plusieurs threads "Save" est actuellement en train de s'exécuter 
+		if (fileLocked.containsKey(filename)) {
+			for (Thread retrieve : fileLocked.get(filename)) { // Si un thread
+																// ou plusieurs
+																// threads
+																// "Save" est
+																// actuellement
+																// en train de
+																// s'exécuter
 				try {
-					retrieve.join(); // On attends la fin de ce thread "Save" (avant de continuer l'exécution du retireve) 
+					retrieve.join(); // On attends la fin de ce thread "Save"
+										// (avant de continuer l'exécution du
+										// retireve)
 				} catch (InterruptedException e) {
 					e.printStackTrace();
 				}
 			}
+		}
 		List<byte[]> bLeft = this.leftSlave.subRetrieve(filename);
+		if(bLeft == null) {
+			return null;
+		}
 		List<byte[]> bRight = this.rightSlave.subRetrieve(filename);
 		return getRecomposeByteArray(bLeft, bRight);
 	}
@@ -164,26 +185,28 @@ public class MasterImpl extends UnicastRemoteObject implements Master {
 
 		// Contruction de l'arbre binaire
 		int i = 1;
-		while (((i+1)*2)-2< slave.length) {
+		while (((i + 1) * 2) - 2 < slave.length) {
 			try {
-				slave[i - 1].setLeftSlave(slave[((i+1)*2)-2]);
-				slave[i - 1].setRightSlave(slave[((i+1)*2)-1]);
+				slave[i - 1].setLeftSlave(slave[((i + 1) * 2) - 2]);
+				slave[i - 1].setRightSlave(slave[((i + 1) * 2) - 1]);
 				i++;
 			} catch (RemoteException e) {
 				e.printStackTrace();
 			}
 		}
-		
-		//A supprimer, test de construction de l'arbre
+
+		// A supprimer, test de construction de l'arbre
 		for (int j = 0; j < (slave.length / 2) - 1; j++) {
 			try {
-				System.out.println("Slave" + slave[j].getId() + " has for left leftSlave slave"+ slave[j].getLeftSlave().getId() + " and has for rightSlave slave" + slave[j].getRightSlave().getId());
+				System.out.println(
+						"Slave" + slave[j].getId() + " has for left leftSlave slave" + slave[j].getLeftSlave().getId()
+								+ " and has for rightSlave slave" + slave[j].getRightSlave().getId());
 			} catch (RemoteException e) {
 				e.printStackTrace();
 			}
 		}
 	}
-	
+
 	private byte[] getRecomposeByteArray(List<byte[]> leftList, List<byte[]> rightList) {
 		byte[] res = new byte[0];
 		for (int i = 0; i < leftList.size(); i++) {
@@ -192,10 +215,10 @@ public class MasterImpl extends UnicastRemoteObject implements Master {
 		for (int i = 0; i < rightList.size(); i++) {
 			res = concat(res, rightList.get(i));
 		}
-		System.out.println("res size : "+ res.length);
+		System.out.println("res size : " + res.length);
 		return res;
 	}
-	
+
 	private byte[] concat(byte[] b1, byte[] b2) {
 		int b1Len = b1.length;
 		int b2Len = b2.length;
@@ -204,7 +227,7 @@ public class MasterImpl extends UnicastRemoteObject implements Master {
 		System.arraycopy(b2, 0, res, b1Len, b2Len);
 		return res;
 	}
-		
+
 	private List<byte[]> getMultipleByteArray2(byte[] fileContent) {
 		System.out.println("Filecontent length : " + fileContent.length);
 		List<byte[]> res = new ArrayList<byte[]>();
@@ -213,17 +236,17 @@ public class MasterImpl extends UnicastRemoteObject implements Master {
 		int cursor = 0;
 		for (int i = 0; i < this.nbSlave; i++) {
 			byte[] forSlave;
-			if(notInRangeByte == 0) {
+			if (notInRangeByte == 0) {
 				forSlave = new byte[byteArrayLength];
 			} else {
-				forSlave = new byte[byteArrayLength+1];
+				forSlave = new byte[byteArrayLength + 1];
 				notInRangeByte--;
 			}
 			for (int j = 0; j < forSlave.length; j++) {
 				forSlave[j] = fileContent[cursor];
 				cursor++;
 			}
-			res.add(forSlave);			
+			res.add(forSlave);
 		}
 		return res;
 	}
@@ -242,7 +265,7 @@ public class MasterImpl extends UnicastRemoteObject implements Master {
 		}
 		int curseur = 0;
 		for (int i = 0; i < this.nbSlave; i++) {
-			byte[] forSlave = new byte[toDivide.length/nbSlave];
+			byte[] forSlave = new byte[toDivide.length / nbSlave];
 			for (int j = 0; j < forSlave.length; j++) {
 				forSlave[i] = toDivide[curseur];
 				curseur++;
